[DT][CPU] Implement VMVXEncodingLayoutAttr.
The revision implements the VMVX encoding layout attribute for device
code in the new IREE::CPU dialect.

Some additional dialect registeration is needed because it could create
the VMVXEncodingLayoutAttr attribute in the pass pipeline; the dialect
needs to be loaded. The revision mainly refactors the logic to the CPU
dialect, the main changes are:

- Move chooseMatmulTile to IREECPUTypes. We can internalize the method
  once other CPU materialization logic is moved to their own attribute
  implementation.
- Move the enumerateMatmulTilesVMVX method to IREECPUAttrs.cpp. It is
  only used by VMVX implementation. It also updates how we check ukernel
  is enabled or not. There is no HAL::ExecutableTarget attribute
  anymore, so it looks at the dictionary configuration directly.

On the data-tiling encoding materialization pass side, it creates the
VMVXEncodingLayoutAttr attribute (with the original target
configuration).

Note that the target configuration is an optional parameter and it is
expected to be used within pass scope, but not the final IR output.

No additional tests because they are covered by
vmvx_materialize_encoding.mlir.
