+++ compiler/src/iree/compiler/Codegen/Dialect/CPU/IR/BUILD.bazel
@@ -1,0 +1,106 @@
+# Copyright 2024 The IREE Authors
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+load("//build_tools/bazel:build_defs.oss.bzl", "iree_compiler_cc_library", "iree_gentbl_cc_library", "iree_td_library")
+load("//build_tools/bazel:enforce_glob.bzl", "enforce_glob")
+
+package(
+    default_visibility = ["//visibility:public"],
+    features = ["layering_check"],
+    licenses = ["notice"],  # Apache 2.0
+)
+
+exports_files([
+    "IREECPUAttrs.td",
+    "IREECPUDialect.td",
+])
+
+iree_td_library(
+    name = "td_files",
+    srcs = enforce_glob(
+        [
+            "IREECPUAttrs.td",
+            "IREECPUDialect.td",
+        ],
+        include = ["*.td"],
+    ),
+    deps = [
+        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:td_files",
+        "@llvm-project//mlir:DialectUtilsTdFiles",
+        "@llvm-project//mlir:OpBaseTdFiles",
+    ],
+)
+
+iree_compiler_cc_library(
+    name = "IREECPUDialect",
+    srcs = [
+        "IREECPUAttrs.cpp",
+        "IREECPUDialect.cpp",
+        "IREECPUTypes.cpp",
+    ],
+    hdrs = [
+        "IREECPUDialect.h",
+        "IREECPUTypes.h",
+    ],
+    textual_hdrs = [
+        "IREECPUAttrs.cpp.inc",
+        "IREECPUAttrs.h.inc",
+        "IREECPUDialect.cpp.inc",
+        "IREECPUDialect.h.inc",
+    ],
+    deps = [
+        ":IREECPUAttrs",
+        ":IREECPUDialectGen",
+        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
+        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/Utils",
+        "//compiler/src/iree/compiler/Dialect/Encoding/IR",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:AffineDialect",
+        "@llvm-project//mlir:DialectUtils",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:LinalgDialect",
+        "@llvm-project//mlir:Parser",
+        "@llvm-project//mlir:Support",
+        "@llvm-project//mlir:TensorDialect",
+    ],
+)
+
+iree_gentbl_cc_library(
+    name = "IREECPUDialectGen",
+    tbl_outs = [
+        (
+            ["--gen-dialect-decls"],
+            "IREECPUDialect.h.inc",
+        ),
+        (
+            ["--gen-dialect-defs"],
+            "IREECPUDialect.cpp.inc",
+        ),
+    ],
+    tblgen = "@llvm-project//mlir:mlir-tblgen",
+    td_file = "IREECPUDialect.td",
+    deps = [":td_files"],
+)
+
+iree_gentbl_cc_library(
+    name = "IREECPUAttrs",
+    tbl_outs = [
+        (
+            ["--gen-attrdef-decls"],
+            "IREECPUAttrs.h.inc",
+        ),
+        (
+            ["--gen-attrdef-defs"],
+            "IREECPUAttrs.cpp.inc",
+        ),
+    ],
+    tblgen = "@llvm-project//mlir:mlir-tblgen",
+    td_file = "IREECPUAttrs.td",
+    deps = [
+        ":td_files",
+        "@llvm-project//mlir:SCFDeviceMappingInterfacesIncGen",
+    ],
+)
