+++ compiler/src/iree/compiler/Codegen/Dialect/CPU/IR/CMakeLists.txt
@@ -1,0 +1,64 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# compiler/src/iree/compiler/Codegen/Dialect/CPU/IR/BUILD.bazel                #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    IREECPUDialect
+  HDRS
+    "IREECPUDialect.h"
+    "IREECPUTypes.h"
+  TEXTUAL_HDRS
+    "IREECPUAttrs.cpp.inc"
+    "IREECPUAttrs.h.inc"
+    "IREECPUDialect.cpp.inc"
+    "IREECPUDialect.h.inc"
+  SRCS
+    "IREECPUAttrs.cpp"
+    "IREECPUDialect.cpp"
+    "IREECPUTypes.cpp"
+  DEPS
+    ::IREECPUAttrs
+    ::IREECPUDialectGen
+    LLVMSupport
+    MLIRAffineDialect
+    MLIRIR
+    MLIRLinalgDialect
+    MLIRParser
+    MLIRSupport
+    MLIRTensorDialect
+    iree::compiler::Codegen::Dialect::Codegen::IR::IREECodegenDialect
+    iree::compiler::Codegen::Dialect::Codegen::Utils
+    iree::compiler::Dialect::Encoding::IR
+  PUBLIC
+)
+
+iree_tablegen_library(
+  NAME
+    IREECPUDialectGen
+  TD_FILE
+    "IREECPUDialect.td"
+  OUTS
+    --gen-dialect-decls IREECPUDialect.h.inc
+    --gen-dialect-defs IREECPUDialect.cpp.inc
+)
+
+iree_tablegen_library(
+  NAME
+    IREECPUAttrs
+  TD_FILE
+    "IREECPUAttrs.td"
+  OUTS
+    --gen-attrdef-decls IREECPUAttrs.h.inc
+    --gen-attrdef-defs IREECPUAttrs.cpp.inc
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
