+++ compiler/src/iree/compiler/Tools/BUILD.bazel
@@ -34,6 +34,7 @@ iree_compiler_cc_library(
     deps = [
         "//compiler/src/iree/compiler/Bindings/Native/Transforms",
         "//compiler/src/iree/compiler/Bindings/TFLite/Transforms",
+        "//compiler/src/iree/compiler/Codegen/Dialect/CPU/IR:IREECPUDialect",
         "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
         "//compiler/src/iree/compiler/Codegen/Dialect/GPU/IR:IREEGPUDialect",
         "//compiler/src/iree/compiler/Codegen/Dialect/VectorExt/IR:IREEVectorExtDialect",
